requirements.txt
praw==7.8.1
google-generativeai==0.8.4
instagrapi==2.1.3
requests==2.32.3
python-dotenv==1.0.0 